fix: quarantine axios types
