+++ src/cli.ts
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
-
-import {upload, RequestResponse} from '.';
+import {upload} from '.';
 
 const args = process.argv.slice(2);
 const opts = {
@@ -10,7 +9,7 @@ const opts = {
 
 process.stdin.pipe(upload(opts))
     .on('error', console.error)
-    .on('response', (resp: RequestResponse, metadata: {mediaLink: string}) => {
+    .on('response', (resp, metadata) => {
       if (!metadata || !metadata.mediaLink) return;
       console.log('uploaded!');
       console.log(metadata.mediaLink);
